refactor: give process preflight explicit ownership
